Registration and login update
added login and registration case in server with all necessary arguments, only thing to do is figure out message protocol and parameters injection.
Customer now has email and password
Regestered customer now has getters and setters
added more functionality to DataBaseManipulation

+++ hibernate_module/src/main/java/il/cshaifasweng/LogInEntities/Employees/Employee.java
@@ -10,7 +10,7 @@ public abstract class Employee implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
     @Column(name="firstName")
     private String firstName;
     @Column(name="LastName")
@@ -62,11 +62,11 @@ public abstract class Employee implements Serializable {
 
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
